+++ src/components/features/card.tsx
@@ -1,5 +1,5 @@
-import { Spinner } from "@/components/Spinner";
 import { Button } from "@/components/ui/button";
+import { Spinner } from "@/components/ui/spinner";
 
 const FeatureCard = ({
   isLoading,
