+++ src/components/navbar/index.tsx
@@ -1,5 +1,5 @@
+import { UserAccountNav } from "@/components/navbar/user-account-nav";
 import { Button } from "@/components/ui/button";
-import { UserAccountNav } from "@/components/user-account-nav";
 import { useSession } from "next-auth/react";
 import Image from "next/image";
 import Link from "next/link";
