+++ src/components/navbar/user-avatar.tsx
