+++ src/pages/f/[docId].tsx
@@ -3,7 +3,7 @@ import dynamic from "next/dynamic";
 const DynamicDocViewerPage = dynamic(
   (() => {
     if (typeof window !== "undefined") {
-      return import("@/components/ReaderScreen");
+      return import("@/components/workspace");
     }
   }) as any,
   { ssr: false },
